Refactored location of the OE App model
Moved the OE App model into the view model for the installed OE apps. Was only being used in that class to the change made sense
